Inventory Improved along with camera effect on player
The inventory system has been changed from text to an image based system along with extra items with their own png file associated to them along with the ability to put any item from the json items file into the game with its own attributes associated to it. A lot of changes and work was put into this commit.